requirements.txt
numpy>=1.24.0
opencv-python>=4.8.0
PyYAML>=6.0
mss>=9.0.0
pynput>=1.7.6
python-statemachine>=2.1.0
Pillow>=10.0.0
